refactor: split app.tsx logic
