--- client/actions/wasmReader.ts
@@ -1,4 +1,0 @@
-export const startGo = (data: any) => {
-  // @ts-ignore
-  return initProject(data);
-};
