+++ client/src/App.css
@@ -9,7 +9,7 @@
   position: absolute;
   top: 3em;
   right: 1em;
-  opacity: .1;
+  opacity: 0.1;
 }
 
 #github a {
@@ -25,7 +25,7 @@ h1 + a {
   position: absolute;
   top: 3em;
   right: 3em;
-  opacity: .1;
+  opacity: 0.1;
 }
 
 .file-selectors-wrapper {
