+++ client/src/App.tsx
@@ -1,100 +1,38 @@
-import { Fragment, useEffect, useRef, useState } from 'react';
-import './App.css';
-import { startGo } from '../actions/wasmReader';
+import { Fragment, useEffect, useState } from 'react';
+import { openVirtualFilesWrapper, saveVirtualFilesWrapper } from '../utils/goFunctions';
 import { Editor } from './components/Editor';
 import { GrDocumentText as DocumentIcon } from '@react-icons/all-files/gr/GrDocumentText';
 import Repository from './components/Repository';
-import { AiFillGithub as GithubIcon } from '@react-icons/all-files/ai/AiFillGithub';
-import { FaQuestionCircle as QuestionIcon } from '@react-icons/all-files/fa/FaQuestionCircle';
-import ReactTooltip from 'react-tooltip';
-import { useRecoilState, useRecoilValue } from 'recoil';
 
-import { virtualFilesState, virtualBranchState } from './atoms/atoms';
+import { useRecoilValue } from 'recoil';
+
+import { virtualFilesState } from './atoms/atoms';
+import Files from './components/Files';
+import HomepageMisc from './components/HomepageMisc';
+
+import './App.css';
 
 function App() {
-  const [virtualFiles, setVirtualFiles] = useRecoilState(virtualFilesState);
-  const virtualBranch = useRecoilValue(virtualBranchState);
+  const virtualFiles = useRecoilValue(virtualFilesState);
 
-  const [file, setFile] = useState<File>();
-  const [fileContent, setFileContent] = useState<any>();
   const [editorContent, setEditorContent] = useState<string>('');
-  const [virtualFileCreation, setVirtualFileCreation] = useState<string>('');
   const [openFile, setOpenFile] = useState<string>('');
   const [disableAll, setDisableAll] = useState<boolean>(false);
-  const [disableFileCreation, setDisableFileCreation] = useState<boolean>(false);
   const [prevOpenedFiles, setPrevOpenedFiles] = useState<Array<string>>([]);
 
-  const fileCreationInput = useRef(null);
+  const clickOnFileHandler = (virtualFile: never) => {
+    setEditorContent(openVirtualFilesWrapper(virtualFile));
+    setOpenFile(virtualFile);
+    setPrevOpenedFiles((prev: Array<string>) => [...prev, virtualFile]);
+  };
 
   const saveChanges = (content: string) => {
-    console.log(content, openFile);
-    // @ts-ignore
-    saveVirtualFile(
-      JSON.stringify({
-        files: [
-          {
-            name: openFile,
-            content: content,
-          },
-        ],
-      }),
-    );
+    saveVirtualFilesWrapper(openFile, content);
   };
 
   const closeEditor = () => {
     setOpenFile('');
   };
-
-  const startGoWrapper = (filename: string, content: string = `Created at: ${new Date().toLocaleString()}`) => {
-    return startGo(
-      JSON.stringify({
-        files: [
-          {
-            name: filename,
-            content: content,
-          },
-        ],
-      }),
-    );
-  };
-
-
-  useEffect(() => {
-    const fileSelector = document.getElementById('file-selector') as HTMLInputElement;
-    fileSelector.addEventListener('change', (event) => {
-      // @ts-ignore
-      const fileList = event?.target!.files;
-      setFile(fileList[0]);
-    });
-
-  }, []);
-
-  // useEffect(() => {
-  //   // @ts-ignore
-  //     if (window.getDefaultFiles) setVirtualFiles(window.getDefaultFiles().split(" "))
-  // }, [window.getDefaultFiles])
-
-  useEffect(() => {
-    if (file) {
-      const reader = new FileReader();
-      reader.onload = (evt: ProgressEvent<FileReader>) => {
-        setFileContent(evt?.target!.result);
-      };
-      reader.readAsText(file);
-    }
-  }, [file]);
-
-  useEffect(() => {
-    if (file && fileContent) {
-      const created = startGoWrapper(file.name, fileContent);
-      if (created) {
-        setVirtualFiles(created.split(' '));
-        const fileSelector = document.getElementById('file-selector') as HTMLInputElement;
-        fileSelector.type = 'text';
-        fileSelector.type = 'file';
-      }
-    }
-  }, [fileContent]);
 
   useEffect(() => {
     const layer = document.querySelector('#layer') as HTMLDivElement;
@@ -107,81 +45,16 @@ function App() {
     }
   }, [openFile]);
 
-  useEffect(() => {
-    if (!!virtualFileCreation) {
-      setDisableFileCreation(false);
-    } else {
-      setDisableFileCreation(true);
-    }
-  }, [virtualFileCreation]);
-
   return (
     <Fragment>
-      <h1>Harmony ✨</h1>
-      <a style={{ color: 'white' }} data-tip="custom show" data-event="click focus">
-        <QuestionIcon size={19} />
-      </a>
-      <ReactTooltip place="bottom" globalEventOff="click">
-        Create and/or modify local files, on the fly!
-        <br />
-        Create/import multiple files Click a file to edit it
-        <br />
-        and jump between them
-        <br />
-        without losing your changes.
-      </ReactTooltip>
-      <div id="layer"></div>
-      <div id="github">
-        <a href="https://github.com/thomscoder/harmony" target="_blank">
-          <GithubIcon size={20} />
-        </a>
-      </div>
-      <div className="file-selectors-wrapper">
-        <label className="custom-file-upload">
-          <input type="file" id="file-selector" disabled={disableAll || !!virtualBranch === false} />
-          Upload file
-        </label>
-        <p>or</p>
-        <div>
-          <form
-            onSubmit={(e) => {
-              e.preventDefault();
-              const created = startGoWrapper(virtualFileCreation);
-              setVirtualFiles(created.split(' '));
-              (fileCreationInput.current! as HTMLInputElement).value = '';
-            }}
-          >
-            <input
-              disabled={!!virtualBranch === false}
-              ref={fileCreationInput}
-              type="text"
-              id="create-file-by-name"
-              name="create-file-by-name"
-              placeholder="e.g. example.txt"
-              onChange={(e) => {
-                setVirtualFileCreation(e.target.value);
-              }}
-            />
-            <button type="submit" id="create-file-btn" disabled={disableAll || disableFileCreation}>
-              Create file
-            </button>
-          </form>
-        </div>
-      </div>
+      <HomepageMisc />
+      <Files disableAll={disableAll} />
       <div className="App">
         <div className="files-area">
           {virtualFiles.map((virtualFile, index) => {
             return (
               <div key={index} className={`virtual-file-wrapper ${!!prevOpenedFiles.find((f) => f === virtualFile) ? 'modified' : ''}`}>
-                <DocumentIcon
-                  size={60}
-                  onClick={() => {
-                    // @ts-ignore
-                    setEditorContent(openVirtualFile(virtualFile));
-                    setOpenFile(virtualFile);
-                    setPrevOpenedFiles((prev: Array<string>) => [...prev, virtualFile]);
-                  }}
-                />
+                <DocumentIcon size={60} onClick={() => clickOnFileHandler(virtualFile)} />
                 <div key={index} className="file">
                   {virtualFile}
                 </div>
