+++ client/src/components/Repository.tsx
@@ -1,13 +1,13 @@
-import Branches from "./Branches"
-import Commit from "./Commits"
+import Branches from './Branches';
+import Commit from './Commits';
 
 const Repository = () => {
-    return (
-        <>
-        <Commit />
-        <Branches />
-        </>
-    )
-}
+  return (
+    <>
+      <Commit />
+      <Branches />
+    </>
+  );
+};
 
-export default Repository
+export default Repository;
