+++ client/src/main.tsx
@@ -4,7 +4,6 @@ import App from './App';
 import './index.css';
 import { RecoilRoot } from 'recoil';
 
-
 ReactDOM.createRoot(document.getElementById('root') as HTMLElement).render(
   <React.StrictMode>
     <RecoilRoot>
