+++ server.js
@@ -4,10 +4,12 @@ var host = process.env.HOST || '0.0.0.0';
 var port = process.env.PORT || 5000;
 
 var cors_proxy = require('cors-anywhere');
-cors_proxy.createServer({
+cors_proxy
+  .createServer({
     originWhitelist: [], // Allow all origins
     requireHeader: ['origin', 'x-requested-with'],
-    removeHeaders: ['cookie', 'cookie2']
-}).listen(port, host, function() {
+    removeHeaders: ['cookie', 'cookie2'],
+  })
+  .listen(port, host, function () {
     console.log('Running CORS Anywhere on ' + host + ':' + port);
-});
+  });
